+++ common/src/main/java/net/mehvahdjukaar/polytone/mixins/AnimatedTextureMixin.java
@@ -2,6 +2,7 @@ package net.mehvahdjukaar.polytone.mixins;
 
 import net.mehvahdjukaar.polytone.texture.DayTimeTexture;
 import net.mehvahdjukaar.polytone.texture.DayTimeTextureTicker;
+import net.minecraft.SharedConstants;
 import net.minecraft.client.renderer.texture.SpriteContents;
 import net.minecraft.client.renderer.texture.SpriteTicker;
 import org.spongepowered.asm.mixin.Final;
@@ -22,22 +23,35 @@ public abstract class AnimatedTextureMixin implements DayTimeTexture {
     @Final
     SpriteContents field_28469;
     @Unique
-    private boolean polytone$usesWorldTime = false;
+    private Mode polytone$mode = Mode.GAME_TIME;
+    @Unique
+    private int polytone$dayDuration = SharedConstants.TICKS_PER_GAME_DAY;
 
     @Override
-    public boolean polytone$usesDayTime() {
-        return polytone$usesWorldTime;
+    public Mode polytone$getMode() {
+        return polytone$mode;
     }
 
     @Override
-    public void polytone$setUsesDayTime(boolean usesWorldTime) {
-        this.polytone$usesWorldTime = usesWorldTime;
+    public void polytone$setMode(Mode mode) {
+        this.polytone$mode = mode;
+    }
+
+    @Override
+    public void polytone$setDayDuration(int duration) {
+        this.polytone$dayDuration = duration;
+    }
+
+    @Override
+    public int polytone$getDayDuration() {
+        return polytone$dayDuration;
     }
 
     @Inject(method = "createTicker", at = @At("HEAD"), cancellable = true)
     public void polytone$modifyTicker(CallbackInfoReturnable<SpriteTicker> cir) {
-        if (polytone$usesWorldTime) {
-            var t = new DayTimeTextureTicker((SpriteContents.AnimatedTexture) (Object) this, this.field_28469, this.interpolateFrames);
+        if (polytone$mode != Mode.GAME_TIME) {
+            var t = new DayTimeTextureTicker((SpriteContents.AnimatedTexture) (Object) this, this.field_28469,
+                    this.interpolateFrames, this.polytone$dayDuration, this.polytone$mode);
             cir.setReturnValue(t);
         }
     }